Add files via upload
This EDA and visualization project was done on the US data set which is a countrywide car accident dataset, which covers 49 states of the USA. The accident data are collected from February 2016 to Dec 2021, using multiple APIs that provide streaming traffic incident (or event) data. These APIs broadcast traffic data captured by a variety of entities, such as the US and state departments of transportation, law enforcement agencies, traffic cameras, and traffic sensors within the road-networks. Currently, there are about 2.8 million accident records in this dataset.

STEPS I FOLLOWED:

1. Dataset is from Kaggle
2. Perform data preparation & cleaning using Pandas & Numpy
3. Perform exploratory analysis & visualization using Matplotlib & Seaborn
4. Asking & answer questions about the data
5. Summarizing & write a conclusion

FINDINGS:

1. There is no data for the New York in the data set
2. Miami top the list with city having severity of 4
3. Los angels top the list for the city with highest number of accident for visibility < 10 and also < 5
4. only 4% of the cities has the accident numbers > 1000
5. At 17:00 hours the maximum accidents occur with Miami toping the list.
6. Most of the accident occurs during the 3 PM to 6 PM.